+++ Sources/SwiftUIKit/Pickers/SimplePicker/SimplePickerScreen.swift
@@ -18,9 +18,11 @@ import SwiftUI
  e.g. by wrapping a picker in a `List` on iOS, in a vertical
  `ScrollView` on tvOS etc.
  
- Future versions of `SwiftUIKit` may add templates that make
- this easier, but for now you can take a look at the preview
- section below.
+ You can find screen templates in `Templates`. The templates
+ can be used to quickly setup pickers, e.g. for tvOS and iOS.
+ Note that they are just suggestions that intend to make the
+ setup quick. You can always use this view to get a complete
+ control over how the picker looks and behaves.
  */
 public struct SimplePickerScreen<Header: View, Picker: View, Footer: View>: DismissableView {